move copy and exec reports to scripts
