+++ zkevm-circuits/Cargo.toml
@@ -36,6 +36,7 @@ num-bigint = { version = "0.4" }
 subtle = "2.4"
 rand_chacha = "0.3"
 snark-verifier = { git = "https://github.com/privacy-scaling-explorations/snark-verifier", tag = "v2023_04_20", default-features = false, features = ["loader_halo2", "system_halo2"] }
+cli-table = "0.4"
 
 [dev-dependencies]
 bus-mapping = { path = "../bus-mapping", features = ["test"] }
@@ -46,7 +47,6 @@ hex = "0.4.3"
 itertools = "0.10.1"
 mock = { path = "../mock" }
 pretty_assertions = "1.0.0"
-cli-table = "0.4"
 serde = { version = "1.0.130", features = ["derive"] }
 serde_json = "1.0.78"
