+++ zkevm-circuits/src/copy_circuit.rs
@@ -777,45 +777,3 @@ impl<F: Field> SubCircuit<F> for CopyCircuit<F> {
         config.assign_copy_events(layouter, &self.copy_events, self.max_copy_rows, *challenges)
     }
 }
-
-#[cfg(test)]
-mod copy_circuit_stats {
-    use crate::{
-        evm_circuit::step::ExecutionState,
-        stats::{bytecode_prefix_op_big_rws, print_circuit_stats_by_states},
-    };
-
-    /// Prints the stats of Copy circuit per execution state.  See
-    /// `print_circuit_stats_by_states` for more details.
-    ///
-    /// Run with:
-    /// `cargo test -p zkevm-circuits --release --all-features
-    /// get_evm_states_stats -- --nocapture --ignored`
-    #[ignore]
-    #[test]
-    fn get_copy_states_stats() {
-        print_circuit_stats_by_states(
-            |state| {
-                // TODO: Enable CREATE/CREATE2 once they are supported
-                matches!(
-                    state,
-                    ExecutionState::RETURNDATACOPY
-                        | ExecutionState::CODECOPY
-                        | ExecutionState::LOG
-                        | ExecutionState::CALLDATACOPY
-                        | ExecutionState::EXTCODECOPY
-                        | ExecutionState::RETURN_REVERT
-                )
-            },
-            bytecode_prefix_op_big_rws,
-            |block, _, _| {
-                assert!(block.copy_events.len() <= 1);
-                block
-                    .copy_events
-                    .iter()
-                    .map(|c| c.bytes.len() * 2)
-                    .sum::<usize>()
-            },
-        );
-    }
-}
